backend/requirements-base.txt
Django ~= 3.2.9

django-cities-light==3.9.2
django-cors-headers==3.13.0
django-allauth==0.51.0
django-environ==0.4.5
django-extensions==3.2
django-filter==21.1
django-model-utils==4.2.0
django-redis ~= 5.2
django-utils-six==2.0
djangorestframework==3.12
dj-rest-auth==2.2.5
drf-api-logger==1.1.12
drf-spectacular==0.25.1
requests==2.27.1

# WSGI application server
gevent ~= 22.0
greenlet ~= 2.0
gunicorn ~= 20.1
Werkzeug ~= 2.0

# Integrations
boto3 ~= 1.23
raven ~= 6.9
